Concurrent tuning of port to Intel-Phi (a.k.a. MIC) and CrayXK.
Move the '__assumed_aligned' directive used for Intel vectorization
to a macro defined and activated in a single location.  This directive
will be correct for allocated views due to aligned memory allocation.
However, this directive could be problematic for user-constructed
subviews which do not honor the memory alignment conventions.

Add 'print_configuration( std::ostream & )' method to the devices.
For Host the hwloc detection and thread binding is printed.
For Cuda the device names, capabilities, and device selection is printed.

This information is frequently used when testing on a new or changed platform,
such as 'compton' and 'curie'.