Piro: Avoided relying on Jacobian adjoint solve
In some instances, the implementation of the response sensitivities used
to involve a Jacobian adjoint solve, without previously checking if the
latter was enabled.

The new implementation exclusively uses forward solves.
How to support and handle adjoint solve is not decided yet,
and such an aspect is important enough to warrant careful design
in order to give the advanced user complete control.
In the meantime, disallowing it altogether is the cleanest approach.