Work-around for bug#5851.
Internally, Galeri::Xpetra::Parameters holds scalars as doubles in its parameter list.
We just have to be careful that when we ask for scalars from the parameter list, we
pull them off as doubles, not the scalar type of the matrix.

Adding test supplied by Karen Devine to the test suite.