Added a function in Utils to construct rigid body modes for structural acoustics problems (coupled acoustic-elastic wave propagation). Also added another example (Simple-Coupled) and some files which demonstrates its use.
Also, edited the Krylov smoother to use the Ifpack2 additive Schwarz
preconditioner with Ifpack2's ILUT as the local inverse; straight ILUT
does not work in parallel if you feed it a CrsMatrix with just a row
map and off-diagonal blocks exist. This is due to the change from
Kokkos::DefaultSparseOps to Kokkos::AltSparseOps.