Tpetra: Map: Constructors 1 & 2 now only check preconds in debug build.
Map's constructors used to always check preconditions on
numGlobalElements, numLocalElements, and indexBase (where applicable).
These checks use multiple global all-reduces.  They are unnecessary
when the user satisfies certain easy-to-understand preconditions
(for example, that the index base input should be the same on all
processes in the communicator).

I've fixed the first two Map constructors (contiguous uniform and
contiguous nonuniform) so that they only do checks in debug mode
(predicated on HAVE_TPETRA_DEBUG).  I've also optimized the second Map
constructor (contiguous nonuniform) by replacing an all-reduce with a
broadcast.  (The constructor does a sum-scan anyway, so Process
numProcs-1 already has the sum result, which is the global number of
elements.  That process just needs to broadcast it to all the others.)
I also updated the public documentation of the two Map constructors
to tell users how to enable or disable debug checks.

I plan to apply the same changes to the third Map constructor
(noncontiguous).  It's just that the changes I've already finished
should improve performance and pass all the Tpetra tests, so I might
as well push them.