Tpetra: Map: Noncontig ctor now only checks preconds in debug build.
Map's constructors always used to check preconditions on
numGlobalElements, numLocalElements, and indexBase (where applicable).
These checks use one or more global all-reduces.  They are unnecessary
when the user satisfies certain easy-to-understand preconditions (for
example, that the index base input should be the same on all processes
in the communicator).

My previous commit 4bf757c58ffc960822c33d6d6404f88eb9498faf fixed the
first two Map constructors (contiguous uniform and contiguous
nonuniform) so that they only do checks in debug mode (predicated on
HAVE_TPETRA_DEBUG).  That saves at least one global reduction per
constructor invocation.  This commit adds this optimization to the
third (noncontiguous) Map constructor.

This should improve performance for Map construction, as well as for
CrsGraph and CrsMatrix fillComplete (when the method needs to compute
the column Map).

I've also added internal comments suggesting further optimizations
for avoiding a redundant all-reduce in the third Map constructor.
This would require adding a new public Directory constructor and a
new private Map constructor (which would bypass some collectives in
the second Map constructor for computing the directory Map), which
is why I didn't include it in this commit.  I probably won't work on
this for a while, because there are plenty of other opportunities for
optimization.