Provide for batch indenting of C/C++ files based on emacs 'thyra' style
The 'thyra' emacs style is very good for development work to automatically
indent C++ code just by hitting 'tab' or running the 'inden-region' function.
In order to allow for automatic indentation of a source file (perhaps as a git
commit hook), it would be nice to be able to do this auotmatically in batch mode.

That is what this little emacs Lisp function can do when run in emacs batch
mode.  I tried this out and it works.

I ran this on a large Tpetra file as:

$ time emacs -batch Tpetra_CrsMatrix_def.hpp \
  -l <THIS_DIR>/batch-indent-file -f emacs-format-function
...
real	1m4.435s
user	1m4.337s
sys	0m0.080s

This worked to indent the file but it took a very long time.  That is not
good.  This is not a viable way to write any type of git commit hook to fix
the indentation of source files.  At the very best I could use this to fix the
formatting on a bunch of files in batch mode offline.

In any case, I am committing this simple script to the Trilinos git repo in
any case so that I can use it in the future if I need to and to document this
some.