MueLu: added coordinates transfer based on aggregation
Prior to this patch, we used the transpose of the tentative prolongator
to transfer coordinates. It has several issues:
 * We need to do an additional matrix transpose
 * We don't really use the matrix itself, as it is scaled, but rather
   replace the entries with ones
 * It works only for scalar equations
It is still useful for transferring other vectors, but coordinates
deserve better.

The new version implements transfer ussing aggregates. We don't need to
transpose matrix anymore, though we do need to construct an import. It
should work fine for multiple dof/node. In the future, we might want to
store the import on the level, as it may be useful in several places,
for instance, distance laplacian.