Piro: Added generic extensible factory
This class is intended to support building different subtypes for a
given abstract interface (specified as a template parameter) based on
a parameter list passed as an input argument.

The available subtypes are determined at runtime by the user. Every
provider is associated with and identified by a unique string key.

Generic implementations of a few simple providers are included.