Anasazi: Fixed build warning relating to undefined pragma.
Anasazi's Epetra test code was using a special pragma specific to
Microsoft Visual Studio to prevent an unnecessary warning when
building on that platform.  GCC 4.4.6 emits a warning when it
encounters this pragma.  This commit prevents this equally
unnecessary warning by protecting use of the pragma with an
"#ifdef _MSC_VER ... #endif", which should only trigger when
building with Visual Studio.