Fixing timer naming.
The Hierarchy itself does not use mutually exclusive timers.  This is so that the timer
can be renamed to "Level xx" instead of "Hierarchy".  (The latter could be confusing.)
But the timer information that is output doesn't indicate that these timers included
child calls, so I just appended "total", which is consistent with other inclusive timers.