Teuchos: Describable: Added describe() variant that takes std::ostream.
It's troublesome to have to look up each time I want to call describe()
how to create a FancyOStream.  Thus, to save myself (and others) this
trouble, I added a nonvirtual member function describe() to the
Describable interface, which takes an std::ostream& instead of a
FancyOStream&.  This version of describe() always defers to the version
that takes a FancyOStream.  It's not virtual so that subclasses are not
tempted to reimplement it.

I also improved the public documentation of Describable.

Build/Test Cases Summary
Enabled Packages: Teuchos, Kokkos, Tpetra, Ifpack2, Belos, Xpetra, MueLu
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=286,notpassed=0 (14.80 min)
3) SERIAL_RELEASE_SS => passed: passed=267,notpassed=0 (14.83 min)
Other local commits for this build/test group: 06b4ec5