Removing instantations for Kokkos::TPINode (Trilinos #5861, VRI Kanban #2852)
Now that the default Kokkos node type is Kokkos::SerialNode, I might as well
remove the instantation for node type Kokkos::TPINode.  This might break some
Ifpack2 tests (see Trilinos #5861) but it is the right thing to do.  Since
this change, none of the Ifpack2 tests are shown to be failing on any platform
so if this causes problems, I will just disable that Ifpack2 test on the CASL
development machines.