Piro: Added caching provider implementation
The new provider is a hybrid between the sharing provider which holds an
RCP to a prebuilt object and returns it on demand, and the constructing
provider which always build a new object when called.

The caching provider builds a new object when called the first time and
holds a reference to it so that it can be returned on subsequent calls.