Teuchos: PerformanceMonitorBase, TimeMonitor: Array -> map.
PerformanceMonitorBase<T> stores all the counters (usually T =
Teuchos::Time, but I'll say "counters" to be general) created using
getNewCounter().  It used to store them in an Array<RCP<T> >.  This
made lookups inefficient, especially when creating a large number of
counters.

A while back, Jeremie convinced me to change getNewCounter() so that it
would first search for the counter (by name) in the Array<RCP<T> >, and
only create a new counter if it didn't yet exist.  This change removed
the last semantic reason to store the counters in an Array instead of
an std::map.  The only reason that remained was semantic:
PerformanceMonitorBase gave access to the counters through a protected
class method counters(), and protected methods are technically part of
the public interface (even though TimeMonitor is the only class I know
that inherits from it).  This is why I hesitated to change the Array to
an std::map at the time.

I was getting some funny behavior out of TimeMonitor (that inherits
from PerformanceMonitorBase<Time>) and wanted to make sure that timer
lookups were doing the right thing, so I decided to replace the Array
finally with std::map.  This should make timer lookups faster,
especially when creating and using a large number of timers.