Panzer: Cleaned up rcp circular dependencies
This was nasty.  The Teuchos::sublist nonmember function stores a post-destroy on the base sublist.  The physics blocks were storing ther own sublist for cloning on side cell topologies inducing a circular reference to the stkIOResponse object in the observer factories that were set in the parameter list.

This commit removes the observer factories from the input parameter list.  Users must register them directly on the ME factory instead of setting them in the input parameter list.  This is a better way to do things anyway.

This commit additionally reverts the temporary patch fix for this issue:
Revert "Panzer: Fixed RCP circular dependency issue"
This reverts commit 21bc4769720f4af37587a70251977338ada9d285.