Stokhos:  Some work on tree representation of lexicographic basis/Cijk.
This generates a block Cijk tensor.  The hope is this would improve the
performance of sparse mat-vec's by eliminating the sparse indexing.  So
far this hasn't turned out to be the case because the blocks are just
too small.  So far I have only implemented a host version of the
multiply.