Epetra: Fix backwards compatibility issue with 64bit work
A couple classes had their constructors changed to make indexBase a long long
type instead of an int. This caused issues for the backwards compatibility
testing because the tests were still passing indexBase in as an int. To fix this
I had to recreate the old constructors that accepted an int indexBase. There are
a couple usages of indexBase being cast to a long long to prevent more
duplication than was strictly necessary. I have deprecated these constructors so
that they can be removed at the next opportunity to break backwards
compatibility.

There are still two tests that are throwing an error in the backwards
compatibility test which will be addressed in a later commit