Panzer: Method added to keep track of stokhos vector objects
This was required because of the strange way stokhos allocates and
manages its coefficient vectors. In particular, they are released
as RCPs, but the memory is still retained underneath. Thus that memory
is invalid if the original EpetraVectorOrthogPoly goes out of scope.
This change makes it possible to work around that issue, but does not
solve the fundamental problem.