Kokkos: TSQR: Fixed conversion from Scalar to Ordinal.
Mike Parks pointed out that a static_cast from Scalar to Ordinal
might not be defined for certain Scalar and Ordinal types, such as
Scalar=dd_real and Ordinal=int.  I changed the relevant code to use
Teuchos::as instead, which does define this conversion.

Note that this does not fix the problem of missing LAPACK functions
for this case.  We know how to solve that problem, though.

Build/Test Cases Summary
Enabled Packages: Kokkos, Tpetra
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=71,notpassed=0 (4.88 min)
3) SERIAL_RELEASE_SS => passed: passed=59,notpassed=0 (8.23 min)