Adding exclusion of git diff check for generated dependency files for non-git TriBITS projects
SCALE which uses TriBITS uses Hg, not git so this configure setup was always
giving them a configure error/warning.  This change adds an additional check
for the .git directory and then only does the check if the .git directory
exists as well.

This should maintain backward compatibility because it also checks for the git
executable as well.

I manually tested this by doing a verbose configure and then looking for the
new MESSAGE(...) printout that I added.  I manually moved the .git directory
and verified that the MESSAGE(...) printout went away.  This would be hard to
unit test but this is not a critical capability and I think in the future we
might even remove the version control of these generated files.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Disabled Packages: CTrilinos,ForTrilinos,PyTrilinos,Didasko,Mesquite,Phdmesh,Pliris,Claps
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=102,notpassed=0 (1.97 min)
1) SERIAL_RELEASE => passed: passed=102,notpassed=0 (1.16 min)