Teuchos: CommHelpers: Specialized reduceAllAndScatter for int,int.
Tpetra::Distributor initialization (computeReceives) uses
reduceAllAndScatter (a wrapper for MPI_Reduce_scatter) to compute each
process' total number of receives when executing the communication plan.

Known performance issues with the generic implementation of reduceAll (a
wrapper for MPI_Allreduce), as observed by the Zoltan2 developers last
year, induced me to write specializations of reduceAll for Ordinal=int
and various Packet types that bypass the generic implementation.  This
fixed the observed performance issues for reduceAll calls.

Since we want Distributor initialization to be as fast as possible, I've
implemented a bypass for reduceAllAndScatter as well.  Distributor only
needs Packet = int, so I have only added that specialization.  It would
be trivial to add other specializations as needed.