Tpetra: Diagnosed msg tag issue in Distributor.  Added debug output.
For a few weeks, Distributor has had a compile-time option to use
message tags for receives and sends.  The option is off by default.
If enabled, Distributor uses message tags to distinguish different
instances of Distributor, as well as different code paths in
Distributor (computeReceives, three-argument doPosts, and four-
argument doPosts).  (This is useful mainly for debugging, and in
particular for figuring out how to get rid of the barrier between
receives and sends in doPosts.)

It turns out that this code was incorrect.  There was a subtle bug
in the way the code distinguished different instances of Distributor.
Distributor was using a global integer counter, incremented in
Distributor's constructor, to assign an identifying number to each
instance of Distributor.  It then computed the message tag as a
function of this ID and the code path in Distributor.  This only works,
though, if users only ever create instances of Distributor (and by
extension, instances of Map, Import, and Export) with the same
communicator.  Matrix Market input and output for Tpetra::Map was
creating an object (that used Distributor) over a different
communicator, namely a SerialComm on Process 0 alone.  That meant that
Distributor instance IDs on Process 0 were one off from the other
processes' instance IDs, meaning in turn that message tags were off.
This caused deadlock, since message tags weren't matching when they
should have matched.

This code was off by default, so no worries about existing results.
However, I would still like to use message tags as a debugging tool.
Thus, I've made message tags independent of the Distributor instance
by default, but left in the path dependency.

It would be possible to assign a unique ID to each Distributor, by
doing an all-reduce in Distributor's constructor to pick the max of
all instance counters' values over all processes in the Distributor's
communicator.  I have not implemented this yet.

In tracing this issue down, I had to add a lot of optional debugging
output to Distributor, Import, and Export.  I left this code in place,
in order to help us figure out how to take out the barrier between
receives and sends in Distributor.

Build/Test Cases Summary
Enabled Packages: Kokkos, Tpetra, Belos, Ifpack2, Xpetra, MueLu
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=192,notpassed=0 (4.62 min)
3) SERIAL_RELEASE_SS => passed: passed=174,notpassed=0 (0.50 min)
Other local commits for this build/test group: 1db68a0, 36864bb