Tpetra: Distributor: Turned off barrier btw recvs & sends by default.
Distributor used to, by default, execute a barrier between posting
receives (in both versions of doPosts) and posting sends.  This barrier
affects scalability and hinders efforts to overlap communication and
computation.  Removing the barrier used to cause deadlock or crashes on
some platforms.  However, perhaps due to some recent changes, I am
unable to manifest deadlock or crashing when removing the barrier, on
platforms where that used to happen.  Eric Phipps also reports a
successful 256-core run without the barrier (and with resulting improved
performance), on a different platform that used to manifest deadlock in
that case.  Thus, this commit turns off the barrier by default.  This
might be a bit bold of me, but it is also a fast way to check whether
any platform can manifest issues without the barrier.  Restoring the
original default behavior is a one-line change.

Build/Test Cases Summary
Enabled Packages: Kokkos, Tpetra, Belos, Ifpack2, Xpetra
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=156,notpassed=0 (8.37 min)
3) SERIAL_RELEASE_SS => passed: passed=138,notpassed=0 (9.50 min)