Tpetra: DistObject,MultiVector: Added new method to report #packets/LID.
I added the protected method constantNumberOfPackets() to DistObject.
It is meant to be called only by DistObject's doTransfer() method.  If
a subclass' implementation returns nonzero, then the instance of that
subclass promises that for the duration of the doTransfer() call, the
number of packets is the same for all LIDs, and that the return value
is this number of packets per LID.

The method has a default implementation that returns zero.  An
implementation could return zero, but still have a constant number of
packets per LID.  The default implementation preserves the current
behavior of doTransfer().  If a nondefault implementation returns
nonzero, doTransfer() will use this information to avoid unnecessary
allocation and / or resizing of arrays.

MultiVector implements this method to return the number of columns in
the MultiVector instance.  This means that doTransfer() need not
allocate or resize DistObject's num{Ex,Im}portPacketsPerLID_ arrays
for a MultiVector.  This, along with a check for correct size of the
arrays in MultiVector's packAndPrepare() implementation, means that an
Import or Export of a MultiVector should only need to allocate buffers
once per instance, even for repeated Import or Export operations.

Build/Test Cases Summary
Enabled Packages: Tpetra, Belos, Ifpack2, Xpetra, MueLu
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=172,notpassed=0 (5.89 min)
3) SERIAL_RELEASE_SS => passed: passed=156,notpassed=0 (1.83 min)
4) MPI_RELEASE_SS => passed: passed=172,notpassed=0 (15.67 min)